Ticket #307: Bugfix Socket layer: assertion fails on socket connection failure
The problem was not visible before Ticket #330.
On socket connection failure, the socket network manager was generating a FAILURE event for socket event manager but state of the socket was not changed.
It seems that if the socket is used again for new network events before being closed, a socket write network event was triggered and leading to interpret it as a connection success (immediate state changed by network manager).

The bugfix consists to add a new intermediary state of CONNECTING_FAILURE set by the socket network manager in this case to avoid to check new network events on this socket before the socket event manager treats the failure and definitively close the socket.