Ticket #307: SC state mgr: avoid recursive events triggered on socket failure (server side)
In case of socket failure detected by server, the SC shall be closed immediately since we cannot send an ERR message on socket.
The server was trying to send an ERR message which was generating a new socket failure (in a recursive way).
Fix: the function to close the SC was already providing the option but was misused, change parameter value to indicate socket already closed