Ticket #421: Keep asserts enabled in release builds
CMake adds by default the NDEBUG define when compiling in Release or
RelWithDebInfo mode. Disabling asserts leads to a number of warnings
mostly because of unused variables, but sometimes because of trickier
reasons (like switch cases having default: assert(false) ).

For now we keep the asserts in all compilation modes, since none of
them should have a high performance cost (and since our deliveries are
anyway compiled in... Debug mode).