Ticket #455: Almost-fix concurrent memory access in asym crypto
When looking at the sources, the `*mbedtls_pk_rsa()` only copies the structures containing the big ints.

This shallow copy works, because the modified big ints (`rsa.Vi` and `rsa.Vf`) are always empty (`Vi.p` is a null pointer), so the copy contains a new buffer afterwards.
However, this would lead to memory leaks, and if the structure is used beforehand, the fix won't work anymore.