Ticket #459: OPC UA types encoder fix: headers shall be encoded separately even when no field present
Messages with no additional fields than the headers were still including
encoding of the request/response header in the body encoder/decoder.
It is not the case anymore, those encoders/decoders are now empty and
return OK without decoding anything to keep the actual mechanism
without impact on messages treatment.

Signed-off-by: monfort <vincent.monfort@systerel.fr>