Ticket #509: user server-side: fixing user comparison in ActivateSession
Since implementation of user on server-side, the user comparison (now pointers) was always failing in ActivateSession service.
It means the reactivation of a session with the same user was always failing when trying to activate the session on a new secure channel.
Now use a C implementation to compare users correctly.