Ticket #549: Sockets performance issue: quick fix to improve performances
Moving the actual data reading on socket in the loop managing the socket network events.
It will fix this particular issue of performance but could still lead to wait for next cycle in others cases (new connection to a server, accepting connection from a server, adding a socket to write set).