Ticket #565: Generate encodeable types
We now generate our own version of the structures and of the table of
known encodeable types.

We temporarily generate in a new file which is included in the original
file.  In the end, the two files will be merged into one generated file.

The original structures and table have been removed from the
sopc_types.c file.