Ticket #565: Put back everything together
Now that the code is fully functional, we go back to the original
situation where the `sopc_types.c` file is fully generated.

Note: One needs to run clang-format on the generated file.  It would be
to cumbersome to match the working of clang-format in the generation
script.

No functional change.