Ticket #547: Minor change: timeoutHint verification is not required to use request timestamp
We used the request timestamp to check the elapsed time since request was sent in order to evaluate the timeoutHint of publihs requets.
Since it is not required by the OPCUA spec part 4, we should not use the request timestamp and only use the current local time as reference for timeout.
It avoids clock synchronization issue between client and server.

OPC UA spec part 4:
- request timestamp definition:
  "The time the Client sent the request. The parameter is only used for diagnostic and logging purposes in the server."
  => i.e.: the server shall not use it for a functional behavior (e.g.: compute publish request timeout ...)
- request timeoutHint definition:
  "For a Server this timeout is only a hint and can be used to cancel long running operations to free resources."
  "The Server shall check the timeoutHint parameter of a Publish request before processing a Publish response."