Ticket #547: debug traces: remove traces that are triggered too often by Browse service
A debug trace was indicating the failure when searching if a type is a transitive subtype of another.
In case of Browse service it is normal to fail often since each available reference type on a node are tested, therefore it generates too many trace lines and should be avoided.