Ticket #623: server SC key renewal bugfix: in key renewal transitory case, server uses old tokenId with new key sets
On server side, after sending the OPN renew response. The new security key set and associated tokenId is available as current key set but is not yet activated.
Its activation occurs only after the first client request with the new tokenId (and use of new security key sets associated). Therefore if the server shall send a response before the activation of the new key sets it shall use the old tokenId and security key sets associated.

Before this fix the server was using the old tokenId as expected but the new security key sets were used instead of previous one.
It leads the message signature to be invalid in sign mode and the messsage encryption and signature to be invalid in sign&encrypt mode.