Ticket #621: bugfix: do not ignore message to send encoding failure + identify "TooLarge" case
Encoding failure when sending a message may be totally ignored (no error / abort message sent and connection not closed).
When message is too large to be sent a Request/ResponseTooLarge status shall be returned.
On server side an abort chunk shall be sent at least to notify client the request will not be treated.