Ticket #629: Event handler enhancement: use a uintptr_t instead of void* for params
It is only guaranteed by C99 that conversion void* => uintptr_t => void* is valid but not uintptr_t => void* => uintptr_t.
Using uintptr_t instead of void* therefore allows to store either an uintptr_t or a void*, whereas using a void* does not allow to store a uintptr_t.