Ticket #649: delivery script: change tag archive prefix directory name to comply with gitlab
The tag archive created by script has a different md5 checksum than the one create by gitlab on same tag.
This is due to the --prefix option definig the root directory name which was S2OPC-<tag_name> on gitlab instead of <tag_name>.
Fix the directoy name to avoid issue next time. But also indicate to keep the archive of gitlab as tag archive reference to be sure md5sum is the same