Ticket #681: BuiltIn types: extends LocalizedText structure to record multiple localized texts in it
LocalizedText type represents a single LocalizedText on the wire (encoding/decoding).
But in the address space nodes LocalizedText fields shall be able to contain a set of localized text instead of a single one.
Since both messages and nodes types are generated from the foundation schemas we have to share the same type to hold both single and multiple localized texts.