Ticket #681: Read/Write services: manage supported/preferred locales on LocalizedText (except IndexRange on write)
Read service: support preferred locales of the session when reading a LocalizedText as single value, array (including with index range), matrix
Write service: check supported locales of the session when writing a LocalizedText as single value, array or matrix. But index range is not managed, which means localized text overwritten with single locale in this case.