Ticket #690: B model: change strategy to support +25% of max nb SCs
In case we reach the maximum number of simultaneous SCs, we keep +25% SCs slot to establish new SCs and try to remove old one.
If no old one can be closed to accept a new one, the new one is closed.
Thus maximum number of simultaneous SCs is in fact the one configured, +25% slots are only use as temporary buffer.