Ticket #690: UACTT tests: fix known bug files for DoS tests
We should remove all DoS attack from known bug since we fixed the issue in S2OPC.
But on CI of gitlab, the UACTT client does not close the socket just after the CLO message and fails because the server does.
Whereas this behavior is correct and the SC auto-closing mechanism works well, the test fails on gitlab. Moreover the test works perfectly when run on local machine