Ticket #743: update build/mingw/check docker to integrate check 0.14 and expat 2.2.9 (with CMake config files)
For mingw docker:
When adding mbedtls/check libraries installed in the standard libraries of mingw, some headers of mingw seem to be included as well.
E.g.: mbedtls headers mbedtls/*.h are installed in /usr/i686-w64-mingw32/include/ which is considered the mbedtls include path since we use "#include mbdetls/<header.h>". When building code dependant on mbedtls this path is included and leads to include mingw headers that are not supposed to be.
It has (at least) for consequence to make fail tests sopc_ieee_check.h verication on float representation.
The current fix consists to install the mbedtls/check libraries in separate paths from mingw standard libraries to avoid including the mingw standard libraries directory to compile.