Ticket #829: PubSub server test: fix configuration to avoid test failure
Some tests are using incoherent security configuration to check subscriber data is not updated in those cases.
Start state might never be obtained because a message is received before retrieving it and Error state already obtained, manage this case.
The Error state shall be obtained when security incoherent state is found by Subscriber due to Publisher network message content. Enforce it is the case.