Ticket #0: Preparing for renode ci
The architecture for the continuous integration on renode will be (first) 1 pub on a card and 1 sub on another card. For this architecture to be done, some configuration files must be modified.
Plus, the ci configuration for gitlab is also modifies to compile s2opc for zephyr and to launch the tests on renode