Ticket #899: Handle value returned from a SOPC_MethodCallFunc_Ptr more gracefully
Move assertion on the result array just after the call to method call pointer callback and make explicit the issue is due to application incorrect behavior.
Note: there is no OPC UA status code available to return in this case and this should not fail silently since the bug is due to application.