Ticket #953: Extract to service_write_1
Extract operations specific to the write service from the address_space
module to the new service_write_1 module:

  - treat_write_request_WriteValues
  - alloc_write_request_responses
  - dealloc_write_request_responses
  - write_WriteResponse_msg_out

The formerly local operation treat_write_request_WriteValue from
address_space_i is now an operation of the address_space module.
I would have liked to make it less dependent on the t_WriteValue_i type,
but it is not currently possible, because this type is also used for
sending notifications upon modification of the address space.