Ticket #956: Client reverse connection fix: server socket shall be associated to SC token
Associate a waiting SC with a server socket connection was not a good idea, socket index was overwritten in case of several server connections.
And since we do not know on server connection if the server is the expected one we should not restrict the number of server connections to the number of waiting connections either.
We have to wait the RHE to know if the server is an expected one or not.
Therefore the concept of secure connection token is added to associate temporarely an SC token and a socket. An associated timeout is created to close the connection if no RHE is received.
Once the RHE is received, if no connection is compatible the SC token is cleared and socket is closed. Otherwise the content of the waiting SC is copied in the SC token to keep the
SC <-> socket association and the waiting SC index is released