Ticket #1029: Remove is_valid_msg_in
This operation is not very useful, it is simpler to return the validity
status directly when decoding the message body.

Note: The operation used to check that the message was not a NULL
pointer and that it contains a valid type (we rely on the decoder to
perform these checks).