Ticket #1002: fix set_range behavior + add support of index range on matrix
When writing a value with an index range, the whole content provided shall be written as indicated by OPC UA specification.
It was not the case until now since we were accepting the end index not to be present in the destination value.
It is still the case when reading value since the OPC UA specification authorizes it.

Add support of IndexRange on multi-dimensional arrays.