Ticket #1106: UACTT Update: add skipped test due to expected config
Skipped: test case incompatible with config necessary for other tests, and expected result is not normalized

When connecting in None for discovery and trying to CreateSession we return BadServiceUnsupported.
It is not expected result by UACTT (BadSecurityPolicyRejected) but specification does not state which status to return.
Therefore it seems appropriate status anyway since service is not supported on discovery endpoint.