Ticket #1106: UACTT test not supported: ignore unspecified behavior test on request TS
UACTT checks behavior on a field that is indicated to be for diagnostic only:
The time the Client sent the request. The parameter is only used for diagnostic and logging purposes in the Server.

Moreover the UACTT code comment previous the verification is pretty clear that this behavior is not expected.
library/ClassBased/UaRequestHeader/5.4-Err-004.js:21:
// if the server doesn't return the error - that is fine