Ticket #1029: Rework client_send_service_request
We now manage the allocation of the output message in the io_dispatcher
module only.  It was previously mixed between this module and
service_mgr.  This makes it easier to check that the message always gets
deallocated when valid.

This change also permitted to see that an IF condition was always TRUE,
and thus remove it.