Ticket #1029: Add forget_msg_out
This new operation is used to indicate that a message is no longer used
as an output message, and that its ownership has been transferred to
another part of the code.

This operartion will be used for asynchronous discovery request messages
from a client that are stored in a queue waiting for a secure channel to
be established.