Ticket #1029: Remove forget_resp_msg_out
This operation was quite strange: it was deallocating the header, but
keeping the message allocated.  It seems much better to me to use
separate operations instead, which makes clearer for the caller what is
happening.

Note: This change was not possible before the introduction of the
forget_msg_out operation.