Ticket #1014: Fix use of a_buffer_out_state
Move preconditions about this variable at the end (as it is the last
output).

Add missing preconditions about this variable.

Add missing postconditions about this variable.

Remove useless assignments to this variable (when it is already
indeterminate in the precondition, there is no need to set it again to
the same value).