Ticket #1062: Simplify postcond of compute_browse_result
In the postcond of the operation, details related to the update of
output variables and parameters are not necessary. In fact, these
details are not required to verify the machine invariant. The operation
compute_rbrowse_result is called by compute_browse_result_from_source,
which has no require on these details.
The details related to the update of output variables and parameters can
be removed from the proof point of view.
Another solution, which is not implemented, is bringing down these details
into a refinement.

Signed-off-by: Laurent Voisin <laurent.voisin@systerel.fr>