Ticket #1067: Cleanup precond of server_validate_async_session_service_resp
Remove commented out properties.

Remove the property stating that an output message is allocated (it is
irrelevant).

However, keep the precondition indicating that we are processing an
output message of the right type to ensure that this operation is called
in the right context (although not needed to implement it).