Ticket #1062: Simplify spec of fill_browse_result_ref
The specification of fill_browse_result_ref gives details of updates of
variables, especially in function of many conditions : reference is
filtered or not, node is unknown or node is known and NodeClass is compatible with NodeClassMask, and also the succes of write operation into output browse variables. These details are not necessary in point of view of proof.
Furthermore, the use of substitution 'CHOICE' in the called operation setall_browse_result_reference_at expresses already the issue of writing into output variables : it is not necessary to duplicate the issue in the calling operation.

The specification of fill_browse_result_ref must be simplified to reduce
the number of generated PO.

Signed-off-by: Laurent Voisin <laurent.voisin@systerel.fr>