Ticket #1062: Split operation treat_one_translate_browse_path
The implementation of treat_one_translate_browse_path is too complicated
to be proved w.r.t its specification (many IF branches and WHILE loop).
In addition, this implementation génerates a great number of PO of which
many are not automatically proved.

One solution is to encapsulate the WHILE loop into a local operation,
named treat_one_translate_browse_path_1.