Ticket #1062: Simplify operation spec
The specification of treat_one_translate_browse_path gives details
related to how variables are updated. These details are not necessary for
its calling operation and generate a great number of POs difficult to
prove.
This commit removes these details. Needed postcond will be added into
the spec.
As the identifier browsePath is used in the operation parameter, the
quantifier browsePath in the invariant d_inv_out, which is
used in the postcond of the operation, is renamed to browse_path.