Ticket #1062: Split operation copy_browsePathResult_to_msg
The implementation of operation copy_browsePathResult_to_msg is too
complicated from a proof point of view : 2 WHILE loops with exit
conditions. A great number of POs are generated, and many of them are not
proved automaticcaly.
One solution is to encapsulate each WHILE loop into a local operation :
copy_browsePathResult_to_msg_1 and copy_browsePathResult_to_msg_2 are
created.