Ticket #1062: Add invariant on in_maxReferencesPerNode
The operation create_continuation_point requires
'p_maxTargetRef <= k_n_BrowseTarget_max' as precond. This operation is
called by compute_browse_result in which p_maxTargetRef takes the value
of in_maxReferencesPerNode.
The predicate 'p_maxTargetRef <= k_n_BrowseTarget_max' is verified when
the browse context is defined ('isBrowseValueContextDefined = TRUE').
As this predicate is required by compute_browse_result, which is not a
local operation, a solution is adding it as an invariant of the B
machine defining the browse_treatment operation.