Ticket #1014: Weaken postcond about req_handle
When the processing fails, the handle is not used by the caller
in io_dispatch_mgr_i so there is no risk in weakening the
post-condition.

What happens in practice, is that in some error cases we remove the
request handle without resetting the output parameter to an
indeterminate value.