Ticket #1062: Rewrite impl of copy_browsePathResult_to_msg
In copy_browsePathResult_to_msg, the copy of remaining path is
proceeded, which is implemented in copy_browsePathResult_to_msg_2, only
if all elements of BrowsePathResult is moved into BrowsePath_Res_X with
success.
This rewrite does not change the functional behavior. It facilitates
proof activity as when calling copy_browsePathResult_to_msg_2, the size
of BrowsePath_Res_Target_Node(browsePath) is known.