Ticket #1062: Update pre/postcond of copy_browsePathResult_to_msg_1
In the postcond of copy_browsePathResult_to_msg_1, 'browsePath : dom(BrowsePath_Res_Target_Node)'
can be removed, as the precond expresses this predicate.

The postcond is updated with the case in which the operation ends with
success. This update is needed to prove the refinement of the calling
operation copy_browsePathResult_to_msg. Therefore, the precond
'size_result = size(BrowsePathResult)' is added.