Ticket #1062: Update pre/postcond of copy_browsePathResult_to_msg_2
In the postcond of copy_browsePathResult_to_msg_2, 'browsePath : dom(BrowsePath_Res_Target_Node)'
can be removed, as the precond expresses this predicate.

The pre/postcond of the operation takes into account the success of
previous called operation copy_browsePathResult_to_msg_1.

The postcond is also updated with the case in which the operation ends with
success. This update is needed to prove the refinement of the calling
operation copy_browsePathResult_to_msg. Therefore, the precond
'nb_max_ref = size(BrowsePathRemainingNodeId)' is added.