Ticket #1014: Fix address space change
Now that we have the add_node service, all the address space can change
when treating a request.  We therefore need to extend the
post-conditions of the local operations that process services.

Note: It might have been better to introduce a cutting machine with
these local operations to abstract away the address space in which we
are not interested at this level of abstraction.