Ticket #1014: Indicate whether client or server
For the operations processing requests and responses, specify in the
precondition whether it is a client or a server channel. This can be
done by looking at the domain of the a_endpoint function, like what is
done by channel_mgr.  The caller (io_dipatcher already performs this
check, so this change does not add new proof obligations elsewhere).

This change reduces the number of POs generated in the
implementation (especially when one could not prove that the channel was
of the right type).