Ticket #1062: Correct inv on BrowsePath_StartingNode
The starting node of a browse path can be undetermined : in the
specification of OPC UA 1.04 table 40, there is no constraint on the
characteristics of the starting node. The implementation of operation
treat_one_translate_browse_path matches the specification of OPC UA : a
check of starting node, realized by check_startingNode, is done.
Therefore, the codomain of BrowsePath_StartingNode must be t_NodeId_i.