Ticket #1062: Correct postcond of treat_browse_request_BrowseValues
In the postcond of treat_browse_request_BrowseValues, the second
'forall' predicate is not correct as the status
e_sc_bad_no_continuation_points is not taken into account.

Moreover, PO generated by the use of the 'forall' predicate are too
complicated to be proved either automatically or by user rules.
Therefore, 'forall' predicates are removed and the postcond is
rewritten. Note that predicates in this rewriting are not equivalent to
the original 'forall' predicates, but they observe the implementation of
the operation.