Ticket #1062: Correct inv related to in_BrowseValue_NodeId
In the machine browse_treatment_context, the invariant
'isBrowseValueContextDefined = TRUE => in_BrowseValue_NodeId : t_NodeId'
is incorrect :
- in_BrowseValue_NodeId is set by setall_browse_value_context with the
parameter p_nodeId. This parameter can be obtained from the input
parameter p_nodeId of set_browse_value_context at the level of
browse_treatment component. This parameter is obtained from the result
of the operation get_BrowsePathSource, which is called in the level
translate_browse_path. As the result of get_BrowsePathSource can be
c_NodeId_indet, the parameter p_nodeId can not be typed with t_NodeId.

- in_BrowseValue_NodeId is used by getall_browse_value_context with the
output parameter p_nodeId. This output is used in the operation
compute_browse_result at the level of browse_treatment, by the input
parameter of the following operations :
  + getall_SourceNode_NbRef : the implementation of this operation shows
  that the typing t_NodeId is not required by calling
  readall_AddressSpace_Node.
  + create_continuation_point : the implementation of this operation
  shows that the typing t_NodeId is not required by realizing the check
  'c_NodeId_indet' in the implementation of
  create_continuation_point_bs.

The invariant must be removed. Related operation precond are also
removed.
Note that in the operation set_browse_value_context at the level
translate_browse_path, the precond 'p_nodeId : t_NodeId' is kept :
this operation is called at the level service_set_view in which a check
'e_sc_bad_node_id_invalid'.