Ticket #1062: Update precond related to relative path element
In the implementation of treat_one_relative_path_element, the calling of
read_RelativePathElt_ReferenceTypeId requires that relativePathElt
belongs to dom(RelativePathElt_ReferenceTypeId). This condition is
verified : relativePathElt is resulted from continue_iter_relativePath.
The codomain of RelativePath_RelativePathElt is a subset of
s_RelativePathElt, and the domain of RelativePathElt_ReferenceTypeId is
s_RelativePathElt.