Ticket #1152: Update typing of 'next' in iterators
Specification of some iterators use 'next' and increase it for each
loop. Variables 'next' being bound (NAT1), a non overflow PO must be
proven at the exit of the last iteration of the loop (<= MAXINT).
This PO is unprovable.

As the 'next' variables are abstract, they can be typed by unbounded
integer set NATURAL1. As 'next' is assigned to an output parameter of
operations 'continue_iter_', the output parameter must be typed with an
implementable type (NAT1) in a before-after predicate.

This commit also removes redundant precond 'nb_browsePathSourceIdx_max_refs > 0'
from the operation 'continue_iter_browsePathSourceIdx'