Ticket #1152: Split operation treat_browse_result_one_source
The implementation of treat_browse_result_one_source makes it difficult
to prove refinement POs (many IF branches in the WHILE loop). The
processing in the loop should be encapsulated in an operation. This new
operation ensures the correctness of writing computed results
(out_BrowseResult_X) to BrowsePathResult, BrowsePathRemainingNodeId and
BrowsePathRemainingIndex for each ReferenceDescription structure.
The specification of treat_browse_result_one_source does not need
checking these details.