Ticket #1188: certificate validation error status: time invalid should only be reported if certificate is trusted
When the certificate is both untrusted and expired. The error to return should be the untrusted error in logs and generic security check failure shall be reported to the client.
Due to the particular mechanism for self-signed certificates, both untrusted and time invalid were reported even when the certificate was trusted leading to not be able to differentiate both cases