Ticket #1188: certificate validation: use PolicyCheckFailed instead of Invalid and make complete validation
In case some Security Policy Check fails during certificate validation steps, the expected CertificatePolicyCheckFailed status was expected to be logged instead of CertificateInvalid (no change for generic status returned to client). Moreover the rest of the certificate validation was not done whereas it is expected to have similar validation check duration.