Ticket #1152: Rewrite translate_browse_path architecture
The translate_browse_path_i component should be rewritten as its size is
too big. One solution is to re-organize operations implementing
breadth-first-search of browse path when computing browse results.
The search uses variables BrowsePathSource, BrowsePathResult and
BrowsePathRemaining, which are initially encapsulted in the component
translate_browse_path_1. The re-organization is done as follows :
- Create the translate_browse_path_result module encapsulating operations
related to access / update of BrowsePathResult and BrowsePathRemaining.
- Create the translate_browse_path_source module encapsulating operations
related to access of BrowsePathSource and some basic operations related
to update of this variable. Not all update BrowsePathSource operations
can be integrated in this module : in fact, the update of BrowsePathSource
depends on BrowsePathResult and vice-versa. The SEES link is not
symmetric. The choice in this re-organisation is that
translate_browse_path_result SEES translate_browse_path_source.

This re-organisation results in modifications:
- Operations in translate_browse_path_1 are dispatched to
translate_browse_path_source_1 and translate_browse_path_result_1.
- Some local operations in translate_browse_path_i are dispatched to
translate_browse_path_source and translate_browse_path_result. The
component translate_browse_path_i keeps the sequencing of process of
BrowsePathSource and BrowsePathResult and some update operations for
BrowsePathSource.