Ticket #1147: remove unused types detecting placeholders
Placeholders are used as object examples inthe confnguration file, these
special objects shall not count as actual references to the types, they
should be removed along with the type, of no other references are found.

An object or variable is considered a placeholder if it has a reference
of type 'HasModellingRule'.