Ticket #1274: SC layer: ensure internal send message is not interupted by extern event
Since sending an OPC UA message is done in 2 times: 1 external event received, 1 internal event triggered and then 1 internal event received.
It is possible that a SC_SERVICE_SND_MSG event is treated but another external event close the secure channel before the associated INT_SC_SND_MSG_CHUNKS internal event generated is treated.
But on an external point of view we want to be able to rely on the external event enqueued to ensure a message is sent prior to SC is closed (ApplyChanges method call request expecteds response to be sent).

As a consequence, use the SOPC_SecureChannels_EnqueueInternalEventAsNext to ensure internal event is treated in priority for sending messages.
Note: it cannot preempt an internal close event channel since it is emitted by the same thread