Ticket #1275: Server clear config: add an assertion to check server is stopped on clear
Make failure explicit if server is cleared without being stopped prior to it.
Future mechanisms will make it easier to comply with by making StopServer waiting for actual closure of EPs in StartServer use case.
For now stopped callback shall be awaited by application in this case to guarantee it