Ticket #1147: define namespace prefix for UA Types
The elements String and ListOfString used in ServerArray and
NamespaceArray belong to the UA Types namespace, which might not be
defined in the NS0 files.
It must be added when merging, else Python serialization produces a
namespace prefix 'ns1'. This namespace cannot be parsed by the Python
parser as it is considered 'for internal use only' !

Declare the UA Types namespace explicitly in the first merged file, with
a 'uax' prefix.

Add a test to verify that calling a merge, then calling it again on the
result with no arguments, produces an identical file and adds the
missing UA Types namespace declaration, with no error.