Ticket #1152: Update loop variant and variable init values
In translate_browse_path_result_i and translate_browse_path_source_i,
the loop variants in operations free_BrowsePathResult,
free_BrowsePathRemaining and free_BrowsePathSource are '0'. They shall
be corrected.

In these operations, the local variable l_index is not initialized
before entering into loops. This makes it impossible to prove loop
invariant related to this variable. An initialization to '0' of this
variable (before entering into loops) corrects the problem and has no
functional impact on the model or the generated code.