Ticket #1152: Rewrite specification of create_continuation_point
In components browse_treatment_continuation_points and
browse_treatment_continuation_points_bs, the specifications of the
operation create_continuation_point uses simultaneous substitution to
evaluate the outputs 'bres' and 'p_ContinuationPointId'. The expression
used to evaluate 'bres' is reused to evaluate 'p_ContinuationPointId' in
an IF condition. These specifications should be rewritten to facilitate
automatic proof : the assignment of 'bres' is moved into the IF
substitution.

The expression 'dom(session_ContinuationPoint~)' is also replaced by
'ran(session_ContinuationPoint)' which is more natural.