Ticket #1152: Correct post-cond of operations in service_set_view_i
In the spec of op treat_browse_next_request_ReleaseContinuationPoint_1,
the post-cond is not sufficient because it does not allow to prove the
preservation of the typing of ResponseBrowseNext_BrowseStatus (invariant
of imported machine). One solution is to add the typing (partial
function) of ResponseBrowseNext_BrowseStatus into the post-cond.

The same kind of issue is identified for op treat_browse_request_BrowseValue_1
and treat_browse_next_request_BrowseContinuationPoint_1.
The same kind of correction is therefore applied to post-cond of these operations.