Ticket #1152: Correct init value of continuation points
Variables modeling continuation points are s_ContinuationPoints,
id_ContinuationPoint and ContinuationPoint_session. In the
INITIALISATION and in operation continuation_points_UNINITIALISATION,
these variables are set to "{}" using "Becomes Element of" substitution.
This is not correct. One of the solution is replacing "Becomes Element
of" by "Becomes Equal".  There is a subtle case for basic machine where
such a change hits a bug in the b2c translator. We work around it by
using the most general kind of substitution.