Ticket #1363: Correct precond related to p_referenceType
In the component browse_treament_continuation_points, the precond of
create_continuation_point "p_referenceType : t_nodeId" is incorrect.
Indeed, the reference type is only defined for nodes of classes
"Object" or "Variable". This precond shall be removed.  It is also
removed in the precond down the callchain to the basic machine.

In this commit, postcond related to p_continuationPointId is also
rewritten for the sake of clarity (operations getall_continuation_point
and release_continuation_point).