Ticket #1363: Rewriting for clarity
Status of browse results are specified by isBrowseResultAllocated and
isBrowseResultComputed. These variables are used and/or assessed in
different level in the B model hierarchy of view service. It would be
better to seperate them from browse result (var "out_BrowseResult_xxx")
in definition clause for the sake of clarity and later update facility.