Ticket #1363: Unmount variables from service_set_view
In service_set_view, declaration of variables "isBrowseValueContextDefined"
and "isBrowseResultComputed" are not necessary. In deed, these variables
always take the value "FALSE" in this component. Consequently, invariant
related to these two variables are reported to service_set_view_i, in
which local operations require precondition related to these two
variables.