Ticket #1147: remove_orphaned fixes: only Root node need to be retained but we should enforce ModellingRules nodes are retained in degraded cases
We keep all the ModellingRules children in the Server node if it is present which is mandatory.
Thus we should not need to retain those nodes specifically, but it is safest to enforce it even in case of Server node absence.

Therefore the place holder of Mandatory and Optional node should also be added to retained node to be complete on available ModellingRules.