Ticket #673: DataType resolution: manage dynamic resolution when type encoder unknown
When an ExtensionObject contains data for an unknown EncodeableType, the DataType can be resolved dynamically using the address space content.
It is expected there is only 1 HasEncoding reference to follow from a 'Default Binary' object node to the corresponding DataType.
Adding this dynamic resolution allow the server to typecheck written values with unknown encoder (extension object value is stotred with ByteString encoded value content).