Ticket #1363: Correct op treat_browse_result_one_source_1
In the current implementation of treat_browse_result_one_source_1, the
browse result node Id is added to BrowsePathResult if its browse name
matches the target name. This leads to unprovable PO of preservetion of
imported machine invariant ran(BrowsePathResult) <: dom(conv_ExpandedNodeId_NodeId).

As the current B model does not deal with the case when the browse name
is out of date (cf. Table 167 of OPC UA Part 4 : local Server might not
continuously monitor the remote Server for changes), the above check of the
target name against the browse name is sufficient in the point of view
of protocol implementation. However, the operation shall be corrected in
order to :
 - Cover the specification of OPC UA even if the out of date of browse
   name has not been implemented yet,
 - Make the above PO provable.

The correction includes :
 - first of all, check if the server is local (expanded node Id can be converted
 into node Id),
 - then check if the browse name matches the looking target name.