Ticket #1363: Update interface of treat_browse_result_one_source_1
The current op treat_browse_result_one_source_1 takes a status as input
parameter and evaluates return code status in function of the input one.
For the sake of clarity, this processing should be brought up to the
calling operation treat_browse_result_one_source.

The default value c_StatusCode_indet is set to the return code in case
the node Id is on local server and its browse name does not match the
target name.