Ticket #1363: Update pre-cond of checkAndAdd_BrowsePathResult
This operation takes an input expanded node Id and possibly adds it into
BrowsePathResult. This expanded node Id belongs to
dom(conv_ExpandedNodeId_NodeId) according to the implementation of the
calling operation treat_browse_result_one_source_1. This predicate shall
be added into the pre-cond of checkAndAdd_BrowsePathResult in order to
preserve the invariant of the imported machine translate_browse_path_result_1
(ran(BrowsePathResult) is a subset of dom(conv_ExpandedNodeId_NodeId)).