Ticket #1363: Correct pre/post-cond of treat_browse_result_one_source_1
The precond "size(BrowsePathResult) < k_n_BrowsePathResPerElt_max" and
"size(BrowsePathRemainingNodeId) < k_n_BrowsePathResPerElt_max" of the
operation treat_browse_result_one_source_1 is incorrect, because the
equality can occur. In adition, properties related to sizes of
BrowsePathResult and BrowsePathRemainingNodeId are expressed in the
invariant of imported machine translate_browse_path_result_1. This
invariant is valid as precondition of the local op
treat_browse_result_one_source_1.

Therefore, the above precond shall be removed from
treat_browse_result_one_source_1.

The postcond shall also be updated to preserve properties related to
sizes of BrowsePathResult and BrowsePathRemainingNodeId. Indeed, the
current specification does not allow to preserve these properties in
case the sizes of inputs BrowsePathResult and BrowsePathRemainingNodeId
are equal to k_n_BrowsePathResPerElt_max.